feat(ts): imports for types
