+++ src/types/@fec/remark-a11y-emoji.d.ts
@@ -1,0 +1,1 @@
+declare module '@fec/remark-a11y-emoji'
