+++ src/types/browser-monads-ts.d.ts
@@ -1,0 +1,7 @@
+declare module 'browser-monads-ts' {
+  export const document: Document
+  export const navigator: Navigator
+  export const window: Window & {
+    gtag?: Gtag.Gtag
+  }
+}
