+++ src/types/graphql.d.ts
@@ -1,0 +1,5 @@
+declare module '*.graphql' {
+  import { DocumentNode } from 'graphql'
+
+  export default typeof DocumentNode
+}
