+++ src/types/remark-code-titles.d.ts
@@ -1,0 +1,1 @@
+declare module 'remark-code-titles'
