+++ tsconfig.json
@@ -21,12 +21,6 @@
       "seoConfig": ["./next-seo.config.ts"]
     }
   },
-  "include": [
-    "next-env.d.ts",
-    "additional.d.ts",
-    "node.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ],
+  "include": ["next-env.d.ts", "src/types/**/*.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
 }
